+++ base_test.py
@@ -44,8 +44,8 @@ from create_suite_page import CaseConductorCreateSuitePage
 from manage_suites_page import CaseConductorManageSuitesPage
 from create_run_page import CaseConductorCreateRunPage
 from manage_runs_page import CaseConductorManageRunsPage
-from create_cycle_page import CaseConductorCreateVersionPage
-from manage_cycles_page import CaseConductorManageVersionsPage
+from create_version_page import CaseConductorCreateVersionPage
+from manage_versions_page import CaseConductorManageVersionsPage
 from create_product_page import CaseConductorCreateProductPage
 from manage_products_page import CaseConductorManageProductsPage
